moves navbar to /components
